--- ft_fork.c
@@ -1,3 +1,0 @@
-#include "minishell.h"
-
-
