+++ minishell.h
@@ -46,89 +46,155 @@ typedef struct s_env
 	struct s_env		*next;
 }						t_env;
 
-// typedef struct s_command_node
-//	{
-//    t_command *cmd;
-//    struct s_command_node *next;
-//} t_command;
-
 typedef struct s_data
 {
 	char				**env;
 	t_env				*env_list;
-	// struct s_data	next;
 }						t_data;
 
-/*erro.c*/
-char					*replace_by_exit_code(char *result, int *result_index);
+/*PARSING_C*/
 
-void					test_pipe(t_command *commands, t_env **env_list);
-t_env					*ftlstlst(t_env *lst);
-/*dollar.c*/
-int						check_builtins(t_command *command, t_env **env_list);
+void	parse_argument(const char **input, char *buffer, int *buf_index);
 
-///void					create_env_list(t_data *data);
-void					create_env_list(char **envp, t_env **env_list);
-void					printf_list(t_env *env_list);
-void					clear_cmd(void);
-void					free_env(t_env *command_list);
-/*prompt*/
+void	handle_quotes(const char **input, char *buffer, int *buf_index, char quote_type);
 
-/*utils.*/
-void					print_commands(t_command *command_list);
-void					free_commands(t_command *command_list);
-int						open_quote(const char *line);
-t_command				*init_command(void);
-void					append_command_node(t_command **head,
-							t_command *new_node);
-char					*ft_strcpy(char *dst, const char *src);
-char					*ft_strcat(char *dst, const char *src);
-int						ft_strcmp(const char *s1, const char *s2);
-/*commands*/
-int						exit_cmd(t_command *command);
-void					start_builtins(t_command *command, t_env **env_list);
-int						choose_command(t_command *command, t_env **env_list);
-void					env_cmd(t_env *env);
-void					echo_cmd(char **args);
-void					cd_cmd(t_command *command, t_env *env_list);
-void					pwd_cmd(void);
-t_env					*export_cmd(t_env *env, t_command *command);
-t_env					*export_args(char *arg, t_env *env_list);
-void					unset_cmd(t_command *command, t_env *env);
-/*pipe_management*/
-void					wait_output(t_command *command, t_env **env_list);
-void					wait_input(t_command *command, t_env **env_list);
-void					commands_manager(t_command *commands, t_env **env_list);
-int						exec_command(char *pathname, char **args);
-/*parsing*/
+int parse_arguments(const char **input, t_command *cmd, int *arg_index);
 
-/*parsingutils*/
-void					parse_argument(const char **input, char *buffer,
-							int *buf_index);
-void					parse_redirection(const char **input, t_command *cmd);
-int						check_for_pipe(const char **input);
-char					*get_env_value(const char *input, int *i);
-int						check_quotes(const char **input, char *quote_type);
-void					handle_quotes(const char **input, char *buffer,
-							int *buf_index, char quote_type);
-void					skip_spaces(const char **input);
-void					handle_dollar_sign(const char **input, char *buffer,
-							int *buf_index);
+int parse_command(const char **input, t_command *cmd);
 
-int						clhandle_double_quotes(const char **input, char *buffer, int *buf_index);
-int						handle_single_quotes(const char **input, char *buffer, int *buf_index);
-char					*search_dollar(const char *input);
-// void					print_commands(t_command *command_list);
-int						parse_command_line(const char *input,
-							t_command **command_list);
-int						parse_command(const char **input, t_command *cmd);
+int	parse_command_line(const char *input, t_command **command_list);
 
-/*signal*/
-void					handle_signal(int sig);
-int						signal_handle(void);
+/*PARSING_UTILS_C*/
 
-/*redirect_management.c*/
-void					redirect_input(t_command *command);
-void					redirect_output(t_command *command);
+t_command	*init_command(void);
+
+t_command	*ft_lstlst(t_command *lst);
+
+void	append_command_node(t_command **lst, t_command *new);
+
+void	quoting_choice(bool *double_q, bool *sing_q, int *index, char c);
+
+int	open_quote(const char *line);
+
+/*DOLLAR_C*/
+
+int	check_builtins(t_command *command, t_env **env_list);
+
+void	process_dollar(const char **input, char *buffer, int *buf_index);
+
+void	handle_dollar_sign(const char **input, char *buffer, int *buf_index);
+
+int	handle_dollar(const char *input, int *i, char *result, int *result_index);
+
+char	*search_dollar(const char *input);
+
+/*ENV_VAR_C*/
+
+void	skip_spaces(const char **input);
+
+char	*get_env_value(const char *input, int *i);
+
+/*REDIRECTIONS_C*/
+
+void	handle_input_redirection(const char **input, t_command *cmd);
+
+void	handle_output_redirection(const char **input, t_command *cmd);
+
+void	parse_redirection(const char **input, t_command *cmd);
+
+/*REDIRECTIONS_MANAGEMENT_C*/
+
+void	redirect_input(t_command *command);
+
+void	redirect_output(t_command *command);
+
+void	redirect_management(t_command *command);
+
+/*COMMAND_MANAGEMENT_C*/
+
+void	commands_manager(t_command *commands, t_env **env_list);
+
+int	choose_command(t_command *command, t_env **env_list);
+
+void	start_builtins(t_command *command, t_env **env_list);
+
+/*EXECUTION_H*/
+
+int	check_path(char *pathname);
+
+int	exec_command(char *pathname, char **args);
+
+/*HEREDOC_MANAGEMENT_C*/
+
+void	wait_output(t_command *command, t_env **env_list);
+
+void	wait_input(t_command *command, t_env **env_list);
+
+void	ft_process_wait(t_command *commands);
+
+
+/*BUILTINS_CD_PWD_C*/
+
+void	update_env(t_env *env_list, char *name, char *value);
+
+void	cd_cmd(t_command *command, t_env *env_list);
+
+void	pwd_cmd(void);
+
+/*BUILTINS_CLEAR_EXIT_C*/
+
+void	clear_cmd(void);
+
+int	exit_cmd(t_command *command);
+
+/*BUILTINS_ECHO_C*/
+
+void echo_cmd(char **args);
+
+/*BUILTINS_ENV_C*/
+
+void	create_env_list(char **envp, t_env **env_list);
+
+void	env_cmd(t_env *env_list);
+
+void	unset_cmd(t_command *command, t_env *env_list);
+
+/*BUILTINS_C*/
+
+t_env	*export_args(char *arg, t_env *env_list);
+
+t_env	*export_cmd(t_env *env_list, t_command *command);
+
+/*BUILTINS_EXPORT_UTILS_C*/
+
+int	ft_is_valid(char **arg);
+
+void	sorted_insert(t_env **head_ref, t_env *new_node);
+
+void	sort_env_list(t_env **head_ref);
+
+/*UTILS_C*/
+
+void	printf_list(t_env *env_list);
+
+char	*ft_strcpy(char *dst, const char *src);
+
+char	*ft_strcat(char *dest, const char *src);
+
+int    ft_strcmp(const char *s1, const char *s2);
+
+void	print_commands(t_command *command_list);
+
+void	free_commands(t_command *command_list);
+
+/*SIGNAL_C*/
+
+void	handle_signal(int sig);
+
+int	signal_handle(void);
+
+/*ERROR_C*/
+
+char	*replace_by_exit_code(char *result, int *result_index);
 
 #endif
